@@ -1,26 +1,27 @@
 {
-  "name": "mdx",
+  "name": "test",
   "version": "1.0.0",
   "description": "",
-  "type": "module",
+  "main": "index.js",
   "scripts": {
-    "compile": "node ./compile-mdx.mjs --root ../../ --json --file"
+    "bundle": "node -r esbuild-runner/register bundle.ts"
   },
   "keywords": [],
   "author": "Michael J. Carter <kiliman@gmail.com> (https://kiliman.dev/)",
   "license": "MIT",
   "dependencies": {
-    "@code-hike/mdx": "^0.3.0-next.14",
-    "@ryanflorence/md": "^7.1.0",
+    "@code-hike/mdx": "^0.7.3",
     "commander": "^9.4.0",
     "dotenv": "^16.0.1",
-    "esbuild": "^0.14.8",
+    "esbuild": "^0.15.5",
     "esbuild-runner": "^2.2.1",
-    "mdx-bundler": "^8.0.1",
-    "node-fetch": "^3.1.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "rehype": "^12.0.0",
-    "shiki": "^0.10.1"
+    "mdx-bundler": "^8.1.0",
+    "node-fetch": "^2.6.7",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "shiki": "^0.11.1"
+  },
+  "devDependencies": {
+    "@types/node-fetch": "^2.6.2"
   }
 }
