@@ -1,7 +1,7 @@
+import type { EntryContext } from '@remix-run/cloudflare'
+import { HandleDataRequestFunction } from '@remix-run/cloudflare'
+import { RemixServer } from '@remix-run/react'
 import { renderToString } from 'react-dom/server'
-import type { EntryContext } from "@remix-run/cloudflare";
-import { HandleDataRequestFunction } from "@remix-run/cloudflare";
-import { RemixServer } from "@remix-run/react";
 
 export default function handleRequest(
   request: Request,
