@@ -1,4 +1,4 @@
-import { ActionFunction, json } from "@remix-run/cloudflare";
+import { ActionFunction, json } from '@remix-run/cloudflare'
 declare var CONTENT: KVNamespace
 declare var POST_API_KEY: string
 
@@ -9,6 +9,7 @@ export const action: ActionFunction = async ({ request }) => {
       return new Response(`Unauthorized ${key}`, { status: 401 })
     }
     const data = await request.json()
+    console.log(`posting ${data.slug}...`)
     await CONTENT.put(data.slug, JSON.stringify(data))
     return json({ success: true })
   } catch (e) {
